requirements.txt
fastapi
uvicorn
pandas
python-dotenv
pinecone-client
requests
tiktoken
pydantic
openai
pinecone
